refactor: create a frame of commands to display the list
